requirements.txt
streamlit
langchain
cohere
streamlit-chat
python-dotenv
langchain-cohere